requirements.txt
pandas
pandas-gbq
google-cloud-dataflow
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
google-cloud-texttospeech
google-auth
db-dtypes
pyyaml
futures
